requirements.txt
# =============================================================================
# GAN Synthetic Image Generator - Python Dependencies
# =============================================================================

# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0

# Image processing
Pillow>=10.0.0
matplotlib>=3.7.0

# Training utilities
tqdm>=4.65.0
PyYAML>=6.0
tensorboard>=2.13.0

# Evaluation
scikit-learn>=1.3.0
scipy>=1.11.0

# Deployment - Streamlit
streamlit>=1.28.0

# Deployment - FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
